[Implode] Add a template implementation for Implode decompression
Please note that Implode is not a class as other (de)compression algorithms like Deflate, but a enum. This is done to test one of the possible future changes to SWCompression: change classes (like Deflate) to enums, since their only function is to represent a namespace for decompress/compress functions.